Move pages inside layout and move layout inside features
