+++ lib/features/layout/pages/mettings/mettings_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
-import '../../core/config/texts/text_styles.dart';
+import '../../../../core/config/texts/text_styles.dart';
 
-class ProfilePage extends StatelessWidget {
-  const ProfilePage({
+class MettingsPage extends StatelessWidget {
+  const MettingsPage({
     super.key,
   });
 
@@ -12,7 +12,7 @@ class ProfilePage extends StatelessWidget {
     return Scaffold(
       body: Center(
         child: Text(
-          "الحساب",
+          "الإجتماعات",
           style: TextStyles.font20BlackBold,
         ),
       ),
