+++ lib/features/layout/pages/quizes/quizes_page.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:gap/gap.dart';
 
-import '../../core/config/texts/text_styles.dart';
-import '../../core/helpers/extensions/extensions.dart';
+import '../../../../core/config/texts/text_styles.dart';
+import '../../../../core/helpers/extensions/extensions.dart';
 import 'widgets/cusotm_search_widget.dart';
 import 'widgets/quiz_item_widget.dart';
