+++ lib/features/layout/pages/quizes/widgets/finished_quiz_item_widget.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:gap/gap.dart';
 
-import '../../../core/config/texts/text_styles.dart';
-import '../../../core/config/theme/colors_manager.dart';
+import '../../../../../core/config/texts/text_styles.dart';
+import '../../../../../core/config/theme/colors_manager.dart';
 import 'quiz_info_widget.dart';
 import 'quiz_result_widget.dart';
 import 'quiz_title_and_description.dart';
