+++ lib/features/layout/pages/videos/videos_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../core/config/texts/text_styles.dart';
+import '../../../../core/config/texts/text_styles.dart';
 
 class VideosPage extends StatelessWidget {
   const VideosPage({
